requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
django==5.2.3
django-allauth==65.9.0
django-cleanup==9.0.0
django-environ==0.12.0
django-htmx==1.23.1
idna==3.10
pillow==11.2.1
pip==25.1.1
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
typing-extensions==4.14.0
urllib3==2.5.0
whitenoise==6.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.0
django==5.2.3
django-allauth==65.9.0
django-cleanup==9.0.0
django-environ==0.12.0
django-htmx==1.23.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
pip==25.1.1
psycopg2==2.9.10
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
typing-extensions==4.14.0
urllib3==2.5.0
whitenoise==6.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
cloudinary==1.44.1
dj-database-url==3.0.0
django==5.2.3
django-allauth==65.9.0
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-environ==0.12.0
django-htmx==1.23.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
pip==25.1.1
psycopg2==2.9.10
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing-extensions==4.14.0
urllib3==2.5.0
whitenoise==6.9.0
